Alpha Build 2.0.1 - 001 ( Alpha of Beta Builds 2.0.1 (Beta of 2.0) )
This *IS* a alpha build!
Things some may *NOT* work, and some things *WILL* not work.

For example, the monitor service is a bit messed up in this build, and
most likely will not work.
Along with that, the global shutdown and reboot commands seem to be
missing, and will not work. (You can still shutdown or reboot the bloody
system)

The update command, for sure, is messed up, and has some unknown
effects.

Do *NOT* use this build *UNLESS* you want a little look at some of the
up coming builds.

Alpha Build 2.0.1 - 001 ( Alpha of Beta Builds 2.0.1 (Beta of 2.0) )

In this build the Boot file was reworked and shrunk by a few hundred
lines. Due to this, most of the system has and is being reworked.