Alpha Build 2.0.1 - 003 ( Alpha of Beta Builds 2.0.1 (Beta of 2.0) )
This *IS* a alpha build!
Things some may *NOT* work, and some things *WILL* not work.
Along with that, the global shutdown and reboot commands seem to be
missing, and will not work. (You can still shutdown or reboot the bloody
system)

The update command, for sure, is messed up, and has some unknown
effects.

Do *NOT* use this build *UNLESS* you want a little look at some of the
up coming builds.

Alpha Build 2.0.1 - 002 ( Alpha of Beta Builds 2.0.1 (Beta of 2.0) )

In this build we update CodeX to version A4.5, added some logging
features to the recovery program, and removed "Run ComputerCraft's
shell" in the BSOD to help with exploits.