Alpha Build 2.0.1 - 004 ( Alpha of Beta Builds 2.0.1 (Beta of 2.0) )
This *IS* a alpha build!
Things some may *NOT* work, and some things *WILL* not work.
Along with that, the global shutdown and reboot commands seem to be
missing, and will not work. (You can still shutdown or reboot the bloody
system)

The update command, for sure, is messed up, and has some unknown
effects.

Do *NOT* use this build *UNLESS* you want a little look at some of the
up coming builds.

Alpha Build 2.0.1 - 004 ( Alpha of Beta Builds 2.0.1 (Beta of 2.0) )

Nothing big came about in this build.

Recovery will now ask for a password before showing troubleshooting
options and will disallow the selection of troubleshooting options
without logging in.
When RitoOS boots it no longer deletes the last log file, rather it adds
it to /System/Logs/RitoOS.log then resets the Latest.log file.
The welcome message to the log file now includes the time the log was
reset along with another slight change in the message.
The old .Recover file was removed.
Added OS function service into the Boot.lua file.